Unify shell load-all orchestration and harden stream parsing.
Merge load-all into useShellGameSelection with one mutation path and shared
query invalidation; validate NDJSON events without casts; split load-all API
tests into focused modules under tests/load_all/.
