+++ packages/frontend/src/api/parseLoadAllStreamEvent.test.ts
@@ -89,6 +89,37 @@ describe('parseLoadAllStreamEvent', () => {
     ).toThrow('invalid shape')
   })
 
+  it('throws when progress event has non-integer perspective', () => {
+    expect(() =>
+      parseLoadAllStreamEvent(
+        JSON.stringify({
+          type: 'progress',
+          phase: 'import',
+          perspective: 1.5,
+          perspective_total: 11,
+          turn: 5,
+          turn_total: 111,
+          message: 'Turn 5',
+        })
+      )
+    ).toThrow('perspective')
+  })
+
+  it('throws when complete event result omits game_id', () => {
+    expect(() =>
+      parseLoadAllStreamEvent(
+        JSON.stringify({
+          type: 'complete',
+          result: {
+            is_game_finished: true,
+            turns_written: 1,
+            turns_skipped: 0,
+          },
+        })
+      )
+    ).toThrow('game_id')
+  })
+
   it('throws on invalid JSON', () => {
     expect(() => parseLoadAllStreamEvent('{not json')).toThrow('invalid JSON')
   })
